016 - add service for trainings
