+++ src/app/homepage/homepage.module.ts
@@ -1,12 +1,13 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import {HttpClientModule} from '@angular/common/http';
 
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatSelectModule } from '@angular/material/select';
+import { MatButtonModule } from '@angular/material/button';
 import { HomepageComponent } from './homepage.component';
 import { HomepageRoutingModule } from './homepage-routing.module';
 import { HeaderModule } from '../header/header.module';
-import { MatSelectModule } from '@angular/material/select';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatButtonModule } from '@angular/material/button';
 import { FooterModule } from '../footer/footer.module';
 import {TrainingModule} from './training/training.module';
 
@@ -22,7 +23,8 @@ import {TrainingModule} from './training/training.module';
     MatSelectModule,
     MatFormFieldModule,
     MatButtonModule,
-    TrainingModule
+    TrainingModule,
+    HttpClientModule
   ],
   providers: [],
   exports: [
